Append space to self-contained completion results
Some completion result types can be reliably assumed to be the end
of an input chunk and followed by a space, so insert that space
automatically if appropriate.

eg a ProviderItem is unlikely to have anything except a space
following it, as it is the end of a path, so insert the space with
the path.

Implements #740, which has more details on the idea.